reduce to separate incoming data
