+++ src/Components/BudgetIndex.js
@@ -2,35 +2,31 @@ import { useContext, useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { ContextData } from "./Provider";
 import BudgetIndexDisplay from "./BudgetIndexDisplay";
+import PendingTransactions from "./PendingTransactions";
+import { convertDate, dateObjCompare, updateTotal } from "../ReusableComponents/helperFunctions";
 import barcode from"../assets/barcode.png"
 import './BudgetIndex.css'
-import { convertDate, dateObjCompare } from "../ReusableComponents/helperFunctions";
 
 
 
 function BudgetIndex() {
-    const {axios, API, originalTotal, setOriginalTotal, data, setData, setHomeModal} = useContext(ContextData)
+    const {axios, API, originalTotal, setOriginalTotal, data, setData, setHomeModal, setPending} = useContext(ContextData)
     const navigate = useNavigate()
     const [currentTotal, setCurrentTotal] = useState(originalTotal)
     const [transactionTotal, setTransactionTotal] = useState(0)
 
-    const [pending, setPending] = useState([])
-    
-   function updateTotal(initValue, arr, setFunction, setFunction2) {
-    let sum = 0
-    arr.forEach(({id, amount}) => {
-        if(id) sum += amount
-    })
-    
-    setFunction(sum)
-    setFunction2(initValue + sum)
-   }
-
     useEffect(() => {
        axios.get(`${API}`)
        .then(respJson => {
-        setData(respJson.data)
-        updateTotal(originalTotal, respJson.data, setTransactionTotal, setCurrentTotal)
+        const [pendingArr, transacArr] = respJson.data.reduce((acc, obj) => {
+            const whichArr = dateObjCompare(obj.date) ? 0 : 1
+            acc[whichArr].push(obj)
+            return acc
+        }, [[],[]])
+
+        setPending(pendingArr)
+        setData(transacArr)
+        updateTotal(originalTotal, transacArr, setTransactionTotal, setCurrentTotal)
        })
        .catch(err => navigate("/*")
     )
@@ -39,7 +35,7 @@ function BudgetIndex() {
     return (
         <div className="index">
             <section className="listedTransactions">
-                <h1>Current Balance: ${originalTotal ? currentTotal.toFixed(2): setHomeModal(true)}</h1>
+                <h1>Current Balance: ${currentTotal.toFixed(2)}</h1>
                 <div className="transactionTitles">
                     <p>{""}</p>
                     <p>Date</p>
@@ -49,23 +45,17 @@ function BudgetIndex() {
                 {
                     data.length > 0 && 
                     data.map(({id, date, itemName, amount}) => {
-                       
                         if(id){
-                            if(dateObjCompare(date)){
-                                // setPending([...pending, id])
-                                console.log(id)
-                            }
-                            else{
+                            if(!dateObjCompare(date)){
                                 return <BudgetIndexDisplay
                                 key = {id}
                                 date={convertDate(date)}
                                 itemName={itemName}
                                 amount={amount}
                                 id={id} />
-                            }
-                            
+                            }   
                         }
-                    })
+                    })   
                 }
                 <img src={barcode} alt="barcode" className="barcode" />
             </section>
